requirements.txt
Django==3.0.8
djangorestframework==3.12.2
markdown==3.2.0
django-rest-swagger==2.2.0
gunicorn==20.0.4
whitenoise==5.2.0
